ReportingStructure, Compensation, and Testing
ReportingStructure: Recursively calculates the number of direct reporters for a given employee by their ID

Compensation: Was unsure if the 'employee' field was of type employee or was meant as the employeeId. To prevent storing the same employee object in two different databases, I implemented it as the employeeId that can be used to reference the employee in the other database. The salary and effectiveDate fields are both strings, as the type was not specified for those either.

Testing: Created a test that creates a Compensation, then attempts to get the same compensation and ensures they are the same. A possible additional test would be to try getting an employee that does not exist and verify the behavior.